[ANDROSDK-1961] Move optionSet api logic to network package
