+++ core/src/main/java/org/hisp/dhis/android/core/option/internal/OptionGroupCall.kt
@@ -30,6 +30,7 @@ package org.hisp.dhis.android.core.option.internal
 import org.hisp.dhis.android.core.arch.api.executors.internal.APIDownloader
 import org.hisp.dhis.android.core.arch.call.factories.internal.UidsCallCoroutines
 import org.hisp.dhis.android.core.option.OptionGroup
+import org.hisp.dhis.android.network.optionset.OptionSetFields
 import org.koin.core.annotation.Singleton
 
 @Singleton
