+++ core/src/main/java/org/hisp/dhis/android/network/optionset/OptionSetFields.kt
@@ -25,29 +25,20 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.option.internal
+package org.hisp.dhis.android.network.optionset
 
-import org.hisp.dhis.android.core.arch.api.HttpServiceClient
-import org.hisp.dhis.android.core.arch.api.payload.internal.PayloadJackson
 import org.hisp.dhis.android.core.option.OptionSet
+import org.hisp.dhis.android.core.option.OptionSetTableInfo.Columns
+import org.hisp.dhis.android.network.common.fields.BaseFields
+import org.hisp.dhis.android.network.common.fields.Field
 import org.hisp.dhis.android.network.common.fields.Fields
-import org.hisp.dhis.android.network.common.filters.Filter
-import org.koin.core.annotation.Singleton
 
-@Singleton
-internal class OptionSetService(private val client: HttpServiceClient) {
-    suspend fun optionSets(
-        fields: Fields<OptionSet>,
-        filter: Filter<OptionSet>,
-        paging: Boolean,
-    ): PayloadJackson<OptionSet> {
-        return client.get {
-            url("optionSets")
-            parameters {
-                fields(fields)
-                filter(filter)
-                paging(paging)
-            }
-        }
-    }
+internal object OptionSetFields : BaseFields<OptionSet>() {
+    val uid: Field<OptionSet> = fh.uid()
+
+    val allFields = Fields.from(
+        fh.getIdentifiableFields(),
+        fh.field(Columns.VERSION),
+        fh.field(Columns.VALUE_TYPE),
+    )
 }
