+++ core/src/main/java/org/hisp/dhis/android/network/optionset/OptionSetService.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2024, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,20 +25,28 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.option.internal
+package org.hisp.dhis.android.network.optionset
 
 import org.hisp.dhis.android.core.option.OptionSet
-import org.hisp.dhis.android.core.option.OptionSetTableInfo.Columns
-import org.hisp.dhis.android.network.common.fields.BaseFields
-import org.hisp.dhis.android.network.common.fields.Field
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
 import org.hisp.dhis.android.network.common.fields.Fields
+import org.hisp.dhis.android.network.common.filters.Filter
+import org.koin.core.annotation.Singleton
 
-internal object OptionSetFields : BaseFields<OptionSet>() {
-    val uid: Field<OptionSet> = fh.uid()
-
-    val allFields = Fields.from(
-        fh.getIdentifiableFields(),
-        fh.field(Columns.VERSION),
-        fh.field(Columns.VALUE_TYPE),
-    )
+@Singleton
+internal class OptionSetService(private val client: HttpServiceClientKotlinx) {
+    suspend fun getOptionSets(
+        fields: Fields<OptionSet>,
+        filter: Filter<OptionSet>,
+        paging: Boolean,
+    ): OptionSetPayload {
+        return client.get {
+            url("optionSets")
+            parameters {
+                fields(fields)
+                filter(filter)
+                paging(paging)
+            }
+        }
+    }
 }
